src/functions/websocket_listener/requirements.txt
azure-functions==1.13.0
websockets==11.0.3
azure-storage-blob==12.18.0
azure-identity==1.14.0
requests==2.31.0
python-dotenv==1.0.0
